feat: make Automaton the sole behavior IR (ARN-382)
TLA+ / StateMachine leftovers let verify and evaluate diverge; parse IOA once to Automaton and derive TemperModel and TransitionTable from that value.
