+++ crates/temper-spec/Cargo.toml
@@ -4,7 +4,7 @@ version.workspace = true
 edition.workspace = true
 license.workspace = true
 rust-version.workspace = true
-description = "CSDL + TLA+ specification parser for the Temper framework"
+description = "I/O Automaton TOML + CSDL specification parser for the Temper framework"
 
 [dependencies]
 quick-xml = { workspace = true }
