+++ crates/temper-spec/src/automaton/parser.rs
@@ -1,14 +1,11 @@
 //! Parse I/O Automaton TOML specifications.
 //!
-//! Also provides conversion to the existing TemperModel and TransitionTable
-//! formats, so the verification cascade and runtime work unchanged.
-//!
 //! The hand-rolled TOML parser lives in [`super::toml_parser`] to keep this
-//! module focused on the public API and validation logic.
+//! module focused on the public API and validation logic. Downstream crates
+//! build `TemperModel` and `TransitionTable` from the resulting [`Automaton`].
 
 use super::toml_parser;
 use super::types::*;
-use crate::tlaplus::{Invariant as TlaInvariant, StateMachine, Transition};
 
 /// Errors from parsing an automaton specification.
 #[derive(Debug, thiserror::Error)]
@@ -360,170 +357,6 @@ fn wire_state_timeout_from_states(automaton: &mut Automaton) {
             }
         }
     }
-}
-
-/// Convert an I/O Automaton to the legacy StateMachine format.
-///
-/// This allows the existing verification cascade (Stateright, DST, proptest)
-/// and the TransitionTable builder to work unchanged.
-pub fn to_state_machine(automaton: &Automaton) -> StateMachine {
-    let transitions = automaton
-        .actions
-        .iter()
-        .filter(|a| a.kind != "output") // Output actions don't transition state
-        .map(|a| {
-            let from_states = if a.from.is_empty() {
-                // Input actions are always enabled (I/O automata property)
-                if a.kind == "input" || a.kind.eq_ignore_ascii_case("composite") {
-                    automaton.automaton.states.clone()
-                } else {
-                    vec![]
-                }
-            } else {
-                a.from.clone()
-            };
-
-            Transition {
-                name: a.name.clone(),
-                from_states,
-                to_state: a.to.clone(),
-                guard_expr: format_guards(&a.guard),
-                has_parameters: !a.params.is_empty(),
-                effect_expr: format_effects(&a.effect),
-            }
-        })
-        .collect();
-
-    let invariants = automaton
-        .invariants
-        .iter()
-        .map(|inv| {
-            // Encode `when` states as a trigger prefix so the model builder
-            // can extract trigger_states via extract_trigger_states().
-            let trigger = if inv.when.is_empty() {
-                String::new()
-            } else {
-                let states = inv
-                    .when
-                    .iter()
-                    .map(|s| format!("\"{s}\""))
-                    .collect::<Vec<_>>()
-                    .join(", ");
-                format!("status \\in {{{states}}} => ")
-            };
-            TlaInvariant {
-                name: inv.name.clone(),
-                expr: format!("{trigger}{}", inv.assert),
-            }
-        })
-        .collect();
-
-    StateMachine {
-        module_name: automaton.automaton.name.clone(),
-        states: automaton.automaton.states.clone(),
-        transitions,
-        invariants,
-        liveness_properties: vec![],
-        constants: vec![],
-        variables: automaton.state.iter().map(|s| s.name.clone()).collect(),
-    }
-}
-
-fn format_guards(guards: &[Guard]) -> String {
-    guards
-        .iter()
-        .map(|g| match g {
-            Guard::StateIn { values } => {
-                format!(
-                    "status \\in {{{}}}",
-                    values
-                        .iter()
-                        .map(|s| format!("\"{s}\""))
-                        .collect::<Vec<_>>()
-                        .join(", ")
-                )
-            }
-            Guard::MinCount { var, min } => format!("{var} >= {min}"),
-            Guard::MaxCount { var, max } => format!("{var} < {max}"),
-            Guard::IsTrue { var } => format!("{var} = TRUE"),
-            Guard::IsFalse { var } => format!("{var} = FALSE"),
-            Guard::ListContains { var, value } => format!("{value} \\in {var}"),
-            Guard::ListLengthMin { var, min } => format!("Len({var}) >= {min}"),
-            Guard::CrossEntityState {
-                entity_type,
-                entity_id_source,
-                required_status,
-                forbidden_status,
-                ..
-            } => {
-                let set = |statuses: &[String]| {
-                    statuses
-                        .iter()
-                        .map(|s| format!("\"{s}\""))
-                        .collect::<Vec<_>>()
-                        .join(", ")
-                };
-                let mut conjuncts = Vec::new();
-                if !required_status.is_empty() {
-                    conjuncts.push(format!(
-                        "{entity_type}[{entity_id_source}].status \\in {{{}}}",
-                        set(required_status)
-                    ));
-                }
-                if !forbidden_status.is_empty() {
-                    conjuncts.push(format!(
-                        "{entity_type}[{entity_id_source}].status \\notin {{{}}}",
-                        set(forbidden_status)
-                    ));
-                }
-                if conjuncts.is_empty() {
-                    "TRUE".to_string()
-                } else {
-                    conjuncts.join(" /\\ ")
-                }
-            }
-        })
-        .collect::<Vec<_>>()
-        .join(" /\\ ")
-}
-
-fn format_effects(effects: &[Effect]) -> String {
-    effects
-        .iter()
-        .map(|e| match e {
-            Effect::Increment { var, amount } => match amount {
-                Some(amount) => format!("{var}' = {var} + {amount}"),
-                None => format!("{var}' = {var} + 1"),
-            },
-            Effect::Decrement { var, amount } => match amount {
-                Some(amount) => format!("{var}' = {var} - {amount}"),
-                None => format!("{var}' = {var} - 1"),
-            },
-            Effect::SetCounterFromParam { var, param } => format!("{var}' = {param}"),
-            Effect::SetBool { var, value } => {
-                format!("{var}' = {}", if *value { "TRUE" } else { "FALSE" })
-            }
-            Effect::Emit { event } => format!("Emit(\"{event}\")"),
-            Effect::Trigger { name } => format!("Trigger(\"{name}\")"),
-            Effect::Schedule {
-                action,
-                delay_seconds,
-            } => format!("Schedule(\"{action}\", {delay_seconds})"),
-            Effect::ListAppend { var } => format!("ListAppend({var})"),
-            Effect::ListRemoveAt { var } => format!("ListRemoveAt({var})"),
-            Effect::ScheduleAt { action, field } => {
-                format!("ScheduleAt(\"{action}\", {field})")
-            }
-            Effect::Spawn {
-                entity_type,
-                entity_id_source,
-                ..
-            } => {
-                format!("Spawn({entity_type}, {entity_id_source})")
-            }
-        })
-        .collect::<Vec<_>>()
-        .join(" /\\ ")
 }
 
 fn validate(automaton: &Automaton) -> Result<(), AutomatonParseError> {
