+++ crates/temper-verify/src/cascade/levels.rs
@@ -1,0 +1,239 @@
+//! Per-level runners for [`super::VerificationCascade`].
+//!
+//! Each method takes the single [`TemperModel`](crate::model::TemperModel)
+//! built in [`super::VerificationCascade::run`].
+
+use crate::checker;
+use crate::model::TemperModel;
+use crate::proptest_gen;
+use crate::simulation::{self, SimConfig};
+use crate::smt;
+
+use temper_runtime::scheduler::FaultConfig;
+
+use super::{ActorSimRunner, CascadeLevel, LevelResult, VerificationCascade};
+
+impl VerificationCascade {
+    /// Level 0: SMT symbolic verification.
+    pub(super) fn run_symbolic_verification(&self, model: &TemperModel) -> LevelResult {
+        let result = smt::verify_symbolic_model(model);
+        let passed = result.all_passed;
+
+        let dead_guards: Vec<&str> = result
+            .guard_satisfiability
+            .iter()
+            .filter(|(_, sat)| !sat)
+            .map(|(name, _)| name.as_str())
+            .collect();
+        let non_inductive: Vec<&str> = result
+            .inductive_invariants
+            .iter()
+            .filter(|(_, ind)| !ind)
+            .map(|(name, _)| name.as_str())
+            .collect();
+
+        let summary = if passed {
+            let mut base = format!(
+                "L0 Symbolic PASSED: {} guards satisfiable, {} invariants inductive, {} unreachable",
+                result.guard_satisfiability.len(),
+                result.inductive_invariants.len(),
+                result.unreachable_states.len(),
+            );
+            if result.approximate {
+                base.push_str("; approximate model: ");
+                base.push_str(&result.approximation_notes.join(" | "));
+            }
+            base
+        } else {
+            let mut issues = Vec::new();
+            if !dead_guards.is_empty() {
+                issues.push(format!("dead guards: {}", dead_guards.join(", ")));
+            }
+            if !non_inductive.is_empty() {
+                issues.push(format!(
+                    "non-inductive invariants: {}",
+                    non_inductive.join(", ")
+                ));
+            }
+            if result.approximate {
+                issues.push(format!(
+                    "approximate model: {}",
+                    result.approximation_notes.join(" | ")
+                ));
+            }
+            format!("L0 Symbolic WARNINGS: {}", issues.join("; "))
+        };
+
+        LevelResult {
+            level: CascadeLevel::SymbolicVerification,
+            passed,
+            summary,
+            verification: None,
+            simulation: None,
+            prop_test: None,
+            smt: Some(result),
+        }
+    }
+
+    /// Level 1: Stateright exhaustive model checking.
+    pub(super) fn run_model_check(&self, model: &TemperModel) -> LevelResult {
+        let verification = checker::check_model(model);
+        let passed = verification.all_properties_hold;
+        let summary = if passed {
+            format!(
+                "L1 Model Check PASSED: {} states explored, all properties hold",
+                verification.states_explored,
+            )
+        } else {
+            let mut parts = Vec::new();
+            if !verification.counterexamples.is_empty() {
+                parts.push(format!(
+                    "{} counterexample(s)",
+                    verification.counterexamples.len()
+                ));
+            }
+            if !verification.dead_transitions.is_empty() {
+                parts.push(format!(
+                    "{} dead transition(s): {}",
+                    verification.dead_transitions.len(),
+                    verification.dead_transitions.join(", ")
+                ));
+            }
+            format!(
+                "L1 Model Check FAILED: {} states explored, {}",
+                verification.states_explored,
+                parts.join("; "),
+            )
+        };
+
+        LevelResult {
+            level: CascadeLevel::ModelCheck,
+            passed,
+            summary,
+            verification: Some(verification),
+            simulation: None,
+            prop_test: None,
+            smt: None,
+        }
+    }
+
+    /// Level 2: Deterministic simulation with fault injection.
+    pub(super) fn run_simulation_level(&self, model: &TemperModel) -> LevelResult {
+        let base_config = SimConfig {
+            seed: 1,
+            max_ticks: self.sim_ticks,
+            num_actors: 3,
+            max_actions_per_actor: 20,
+            max_counter: self.max_counter,
+            faults: FaultConfig::light(),
+        };
+
+        let results = simulation::run_multi_seed_simulation(model, &base_config, self.sim_seeds);
+
+        let invariants_ok = results.iter().all(|r| r.all_invariants_held);
+        let liveness_ok = results.iter().all(|r| r.liveness_violations.is_empty());
+        let all_passed = invariants_ok && liveness_ok;
+        let total_transitions: u64 = results.iter().map(|r| r.total_transitions).sum();
+        let total_dropped: u64 = results.iter().map(|r| r.total_dropped).sum();
+        let violations: Vec<_> = results.iter().flat_map(|r| r.violations.clone()).collect();
+        let liveness_violations: Vec<_> = results
+            .iter()
+            .flat_map(|r| r.liveness_violations.clone())
+            .collect();
+
+        let summary = if all_passed {
+            format!(
+                "L2 Simulation PASSED: {} seeds, {} transitions, {} dropped msgs",
+                self.sim_seeds, total_transitions, total_dropped,
+            )
+        } else if !invariants_ok {
+            format!(
+                "L2 Simulation FAILED: {} invariant violation(s) across {} seeds",
+                violations.len(),
+                self.sim_seeds,
+            )
+        } else {
+            format!(
+                "L2 Simulation FAILED: {} liveness violation(s) across {} seeds",
+                liveness_violations.len(),
+                self.sim_seeds,
+            )
+        };
+
+        let representative = results.into_iter().next();
+
+        LevelResult {
+            level: CascadeLevel::Simulation,
+            passed: all_passed,
+            summary,
+            verification: None,
+            simulation: representative,
+            prop_test: None,
+            smt: None,
+        }
+    }
+
+    /// Level 3: Property-based tests with shrinking for minimal counterexamples.
+    pub(super) fn run_prop_tests_level(&self, model: &TemperModel) -> LevelResult {
+        let result = proptest_gen::run_prop_tests_with_shrinking_on_model(
+            model,
+            self.prop_test_cases,
+            self.prop_test_max_steps,
+        );
+        let passed = result.passed;
+
+        let summary = if passed {
+            format!(
+                "L3 Property Tests PASSED: {} cases, {} max steps",
+                result.total_cases, self.prop_test_max_steps,
+            )
+        } else {
+            let failure_desc = result
+                .failure
+                .as_ref()
+                .map(|f| {
+                    format!(
+                        "invariant '{}' violated after {} actions",
+                        f.invariant,
+                        f.action_sequence.len()
+                    )
+                })
+                .unwrap_or_else(|| "unknown failure".to_string());
+            format!("L3 Property Tests FAILED: {}", failure_desc)
+        };
+
+        LevelResult {
+            level: CascadeLevel::PropertyTest,
+            passed,
+            summary,
+            verification: None,
+            simulation: None,
+            prop_test: Some(result),
+            smt: None,
+        }
+    }
+
+    /// Level 2b: Actor simulation with real TransitionTable::evaluate().
+    pub(super) fn run_actor_simulation(&self, runner: &ActorSimRunner) -> LevelResult {
+        let result = runner(self.sim_seeds);
+
+        let summary = if result.all_invariants_held {
+            format!(
+                "L2b Actor Simulation PASSED: {} seeds, {} transitions",
+                result.seeds_tested, result.total_transitions,
+            )
+        } else {
+            format!("L2b Actor Simulation FAILED: {}", result.summary)
+        };
+
+        LevelResult {
+            level: CascadeLevel::ActorSimulation,
+            passed: result.all_invariants_held,
+            summary,
+            verification: None,
+            simulation: None,
+            prop_test: None,
+            smt: None,
+        }
+    }
+}
