+++ crates/temper-verify/src/lib.rs
@@ -21,7 +21,7 @@ pub use checker::{VerificationResult, check_model};
 pub use composite::{CompositePlanError, CompositeVerificationPlan};
 pub use model::{
     InvariantKind, ModelEffect, ModelGuard, ResolvedTransition, TemperModel, TemperModelAction,
-    TemperModelState, build_model_from_ioa,
+    TemperModelState, build_model_from_automaton, build_model_from_ioa,
 };
 pub use paths::{
     PathExtractionConfig, PathExtractionResult, PathStep, ReachablePath, extract_paths,
